mssqlnative: improve efficiency by reducing schema reads
This is done by reading and caching the schema for the requested table
at the initial request, resulting in a single sqlsrv_field_metadata()
call. All subsequent requests use the cached schema information. This
method is copied from the method used in the oci8 driver.

In addition, the following changes have been made:

1. Date fields are handled better, the ReturnDatesAsStrings parameter is
   now passed in the connection, and the system can now distinguish
   between date(91) and datetime(93) fields correctly, and format them
   according to standards. This is done using the 'format' callback for
   the date/time field.

2. The field object is now returned as an ADOFieldObject instead of an
   array, making it compatible with other drivers.

3. Driver now supports sequence creation in SQL Server 2014

Fixes #70