ADOrecordset::getAssoc poor at identifying type of array, see #101
The ADOrecordset::getAssoc method has suffered from reduced reliability
because the original design forced it to try to identify the type of array
(numeric/associative) by looking at the keys of the array. 
This redesign eliminates such requirements in the following way:

  1. Passes the actual fetch mode from the ADOconnection::getAssoc to
     ADOrecordset::getAssoc. 
  2. If ADOrecordset::getAssoc is used as an API, (This method is discouraged
     and may become unavailable in later versions of ADOdb), the method
	 attempts to identify the array type by analyzing the entirety of the
	 array and comparing it against a numeric array of the same length.
  3. Finally, a fetch mode can be forced by sending one directly to the
     ADOrecordset::getAssoc method as a third argument.
   
The definition of numeric array for the purposes of this method, is:
An array whose keys are ALL integer values greater than or equal to zero,
the first key MUST be the value 0, and the incremental value between keys
MUST be 1. This is not necessarily the definition of a numeric array for
PHP purposes, but reflects numeric arrays as returned by database drivers.